got edit chores to refresh the parent dashboard, now working on child dashboard
