+++ chores-app/src/components/ChoreCard.js
@@ -5,7 +5,15 @@ import PropTypes from "prop-types";
 import EditChore from "./EditChore";
 // need to add the functionality to these functions; edit button should open an edit view with populated fields based on the chore it was clicked in
 // the delete button will remove the chore from the database, and a nice to have would be a "confirm delete".
-const ChoreCard = ({ name, price, status, component, choreID, owner }) => {
+const ChoreCard = ({
+  name,
+  price,
+  status,
+  component,
+  choreID,
+  owner,
+  setRerender,
+}) => {
   const [editing, setEditing] = useState(false);
   const editButton = (event) => {
     event.preventDefault();
@@ -40,6 +48,7 @@ const ChoreCard = ({ name, price, status, component, choreID, owner }) => {
         status={status}
         setEditing={setEditing}
         choreID={choreID}
+        setRerender={setRerender}
         owner={owner}
       />
     );
