+++ chores-app/src/context/UserAuthContext.js
@@ -12,6 +12,7 @@ const userAuthContext = createContext();
 // eslint-disable-next-line react/prop-types
 export const UserAuthContextProvider = ({ children }) => {
   const [user, setUser] = useState({});
+  const [chores, setChores] = useState([]);
 
   function logIn(email, password) {
     return signInWithEmailAndPassword(auth, email, password);
@@ -41,7 +42,7 @@ export const UserAuthContextProvider = ({ children }) => {
   return (
     <userAuthContext.Provider
       // eslint-disable-next-line react/jsx-no-constructed-context-values
-      value={{ user, logIn, signUp, logOut }}
+      value={{ user, logIn, signUp, logOut, chores, setChores }}
     >
       {children}
     </userAuthContext.Provider>
